Harden against SAST findings; add signed macOS .pkg packaging pipeline
Security/robustness fixes (from cppcheck + manual security review):
- EMFSerialReader: drain ALL complete frames per read and cap the read buffer
  (untrusted USB input could desync framing and grow the buffer unbounded); add
  a 1s timeout so a missing <GETVER>> reply never wedges EMF parsing
- SessionRecorder: track audio samples as 64-bit and clamp/warn on the 32-bit
  WAV size fields (no more silent overflow/corruption past ~6h); warn instead of
  silently succeeding when session metadata / EMF CSV fail to open
- main.cpp: QLockFile stale time 30s so a crashed instance's lock is reclaimed
  instead of blocking all future launches
- EMFCorrelator: value-initialize CorrelatedEvent (no uninitialized copy)

macOS packaging / signing / notarization:
- resources/entitlements.plist: hardened runtime + microphone + WebEngine JIT
- scripts/fetch-models.sh: download the bundled Whisper models
- scripts/package-macos.sh: build -> bundle all models into Resources -> macdeployqt
  (bundle Qt incl. WebEngine) -> inside-out Developer ID codesign (helper-first
  ordering so the WebEngine seal stays valid) -> Developer ID Installer-signed
  .pkg (productbuild) -> notarytool (keychain profile) -> staple
- MainWindow: resolve bundled models from Contents/Resources/models
- .gitignore: exclude built installers (dist/, *.pkg)

+++ src/session/SessionRecorder.h
@@ -76,7 +76,7 @@ private:
 
     uint32_t m_audioSampleRate{16000};
     bool m_headerWritten{false};
-    uint32_t m_audioSamplesWritten{0};
+    uint64_t m_audioSamplesWritten{0};  // 64-bit: WAV header field is clamped at finalize
     uint32_t m_snippetCount{0};
     QElapsedTimer m_recordTimer;
 };
